+++ btcstaking-tracker/btcslasher/bootstrapping_test.go
@@ -207,7 +207,7 @@ func FuzzSlasher_Bootstrapping(f *testing.F) {
 			AnyTimes()
 
 		mockBTCClient.EXPECT().
-			SendRawTransaction(gomock.Any(), gomock.Eq(true)).
+			SendRawTransactionWithBurnLimit(gomock.Any(), gomock.Eq(true), gomock.Any()).
 			Return(&chainhash.Hash{}, nil).
 			AnyTimes()